requirements.txt
Flask
numpy
joblib
scikit-learn
opencv-python
tensorflow
pillow
transformers
torch